pyproject.toml
[tool.poetry]
name = "my-pic-backend"
version = "0.1.0"
description = ""
authors = ["stygj <stygj59@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "^23.2.1"
uvicorn = "^0.25.0"
fastapi = "^0.109.0"
fastapi-pagination = "^0.12.14"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
greenlet = "^3.0.3"
gunicorn = "^21.2.0"
bcrypt = "^4.1.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
